requirements.txt
Django==5.1.5
djangorestframework==3.15.0
django-cors-headers==4.7.0
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
setuptools>=42.0.0

# REST API dependencies
django-filter==23.3
markdown==3.5
drf-yasg==1.21.7  # For API documentation
djangorestframework-simplejwt==5.3.0  # For JWT authentication



